# api.yaml
openapi: 3.0.0
info:
  title: vcip-api
  version: '1.0'
  description: Сервисы VCIP
  contact:
    name: Shakhmurat Japparov
    email: japparov.s@telecom.kz
servers:
  - url: 'http://localhost:3000/api/vcip-api/v1.0'
    description: test
paths:
  /trace-optical:
    get:
      summary: Получить трассировку по оптической сети
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/traceOptical'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '408':
          description: Request Timeout
        '500':
          description: Internal Server Error
      operationId: get-trace-optical
      description: Получить трассировку по оптической сети
      parameters:
        - $ref: '#/components/parameters/townStateId'
        - $ref: '#/components/parameters/containerSpecId'
        - $ref: '#/components/parameters/containerExternalId'
        - $ref: '#/components/parameters/mountedPortNum'
      tags:
        - Трассировка по оптической сети
  /trace-copper:
    get:
      summary: Получить трассировку по медной сети
      tags:
        - Трассировка по медной сети
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/traceCopper'
      operationId: get-trace-copper
      description: Получить трассировку по медной сети
      parameters:
        - $ref: '#/components/parameters/townStateId'
        - $ref: '#/components/parameters/containerExternalId'
        - $ref: '#/components/parameters/mountedPortNum'
  /building-geo-cords:
    get:
      summary: Получить геокоординаты здания
      tags:
        - Геокоординаты
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Данные ответа
                    required:
                      - pointLon
                      - pointLat
                    properties:
                      pointLon:
                        type: string
                        description: Долгота
                      pointLat:
                        type: string
                        description: Широта
                required:
                  - data
      operationId: get-building-geo-cords
      description: Метод возвращает геокоординаты здания по параметрам
      parameters:
        - schema:
            type: integer
          in: query
          name: townStateId
          description: Населенный пункт_ID
          required: true
        - schema:
            type: integer
          in: query
          name: streetId
          description: Улица_ID
          required: true
        - schema:
            type: integer
          in: query
          name: buildingNum
          description: Номер здания
          required: true
        - schema:
            type: string
          in: query
          name: buildingSubNum
          description: Дробь здания
components:
  schemas:
    traceOptical:
      title: traceOptical
      x-stoplight:
        id: 6ltvouy02oeyv
      type: object
      description: Данные трассировки по оптической сети
      properties:
        data:
          type: object
          required:
            - lineData
          description: Данные ответа
          properties:
            lineData:
              type: string
              description: Строка линейных данных
              example: 'TOWN:727;STATION:40/01; MAN:ECI; OLT:3/00/06/06;ODF:2/08/09/06; LD:ОРШ 223/37: 00/00/1 магистраль/6, 00/00/2 сплиттеры /16, 00/00/2 сплиттеры /18, 00/00/5 распределение/12; ОРКсп 223/37/6/2: 00/00/1х16/11 (Этаж 5); OU:0;ONUPORTNUM:0; (DSL,SIP); 3.542 км'
      required:
        - data
    traceCopper:
      title: traceCopper
      x-stoplight:
        id: ecyxk6rvarql8
      type: object
      properties:
        townStateId:
          type: integer
          example: 2
          description: Населенный пункт_ID
        parentContainerExternalId:
          type: integer
          example: 13
          description: ID родительского физического контейнера во внешней системе
      description: Данные трассировки по медной сети
  parameters:
    townStateId:
      name: townStateId
      in: query
      required: true
      schema:
        type: number
      description: Населенный пункт_ID
    containerSpecId:
      name: containerSpecId
      in: query
      required: true
      schema:
        type: number
      description: Спецификация физического контейнера_ID
    containerExternalId:
      name: containerExternalId
      in: query
      required: true
      schema:
        type: number
      description: ID физического контейнера во внешней системе
    mountedPortNum:
      name: mountedPortNum
      in: query
      required: true
      schema:
        type: string
      description: Номер монтированного порта
tags:
  - name: Трассировка по оптической сети
  - name: Трассировка по медной сети
  - name: Геокоординаты